Unit testing uncertification
