+++ services/app-api/handlers/stateStatus/tests/uncertify.test.ts
@@ -1,4 +1,4 @@
-import { uncertifyReport } from "../uncertify";
+import { updateStateStatus } from "../update";
 import { APIGatewayProxyEvent } from "aws-lambda"; // eslint-disable-line no-unused-vars
 import { testEvent } from "../../../test-util/testEvents";
 import { UserRoles } from "../../../types";
@@ -14,22 +14,18 @@ jest.mock("../../../libs/dynamodb-lib", () => ({
 jest.mock("../../../libs/authorization", () => ({
   isAuthorized: jest.fn().mockReturnValue(true),
   getUserCredentialsFromJwt: jest.fn().mockReturnValue({
-    //   TODO: update this role from ADMIN to APPROVER
     role: UserRoles.ADMIN,
     state: "AL",
   }),
 }));
-
-describe("Test Uncertify Report Handlers", () => {
+describe("Test Uncertify CARTS Report Handler", () => {
   test("uncertify CARTS report", async () => {
     const event: APIGatewayProxyEvent = {
       ...testEvent,
       pathParameters: { year: "2021", state: "AL" },
       body: `{"status": "in_progress", "username": "test user"}`,
     };
-
-    const res = await uncertifyReport(event, null);
-
+    const res = await updateStateStatus(event, null);
     expect(res.statusCode).toBe(200);
     expect(dynamodbLib.update).toBeCalledWith({
       ExpressionAttributeNames: {
@@ -47,7 +43,7 @@ describe("Test Uncertify Report Handlers", () => {
         year: 2021,
       },
       UpdateExpression:
-        "set #username=:username, #lastChanged=:lastChanged, #status=:status",
+        "set #status=:status, #username=:username, #lastChanged=:lastChanged",
       TableName: undefined,
     });
   });
