+++ services/app-api/handlers/stateStatus/update.ts
@@ -44,4 +44,29 @@ export const updateStateStatus = handler(async (event, _context) => {
 
     await dynamoDb.update(updateStateStatusparams);
   }
+
+  /**
+   * Approver uncertifies a CARTS report
+   */
+  // eslint-disable-next-line
+  // TODO: update ADMIN with APPROVER role
+  else if (user.role === UserRoles.ADMIN) {
+    const uncertifyReportParams = {
+      TableName: process.env.stateStatusTableName!,
+      Key: {
+        stateId: state,
+        year: parseInt(year),
+      },
+      ...convertToDynamoExpression(
+        {
+          status: "in_progress",
+          username: username,
+          lastChanged: new Date().toString(),
+        },
+        "post"
+      ),
+    };
+
+    await dynamoDb.update(uncertifyReportParams);
+  }
 });
